Proguarding the car-support-lib.
Updated update-car-api to build a proguard file if it is requested.
Updated the car-support-lib build to use that proguard file when the android.support.car-prebuilt target is built. (this happens in a dist)
cleaned up some @hide's to make the proguarding work.
Added proguard-release.flags which only needs to be updated when the api is updated.

Bug 33217507

Test: built the android.support.car-prebuilt target in a dist and looked at the contents of the resulting aar file.  All classes with the name Embedded have been obfuscated.

Change-Id: Ic82d7afdef77e2a8a5333c153389c9e4cdb96c25