Creating a BluetoothDeviceConnectionPolicy in CarService for Bluetooth Autoconnect.
A policy driven approach to trigger connection attempts to devices that have
previously connected to the vehicle. The policy listens for signals from the car
to decide on when to make the attempt and maintains a list of devices that it
can attempt the connections on for a given profile.

Bug: b/33080250
Test: New feature, testing using Car tests.
Change-Id: I6086c65a934f1721550a1049f5f7d95ae1d5251a