Clarify the roles of SensorHalServiceBase vs. SensorHalService
SensorHalServiceBase is now responsible for communication with Vehicle HAL, speaking in terms of VehiclePropValue/VehiclePropConfig and subscriptions
SensorHalService builds on SensorHalServiceBase to provide CarSensorEvents and dispatching thereof to listeners.

This introduces a base-interface SensorBase to deal with DrivingState and DayNightMode policy objects which want to act as-if they're sensors to CarSensorService,
but don't truly talk to Vehicle HAL (and thus are not SensorHalServiceBase).

The purpose of this change is to create a base layer for future work which wants to take control of Vehicle HAL properties and read/write them, but does not
vend CarSensorEvents to higher levels of the vehicle stack.

Test: build
Change-Id: Ia28044271b8dc2c996555214f66ac5902b685a75