Update EVS app and manager for HAL revisions
The HAL contract, particuarly with respect to opening and closing
devices has changed and the manager and app have been revised
accordingly.  The previous "test" has been removed as it has been
superceeded by the VTS test for the HAL.

NOTE:  This change was originally developed and +2'd on the master
branch as ag/2021133.  However cherry picking to oc-dev is broken, so
this will be submitted directly to oc-dev and auto-merged back to
master.

Test: build locally against updated EVS HAL
Change-Id: I2378d622079a2326dc7d1432adca72bcf112e193