Don't send cached events from CarDiagnosticService
Upon registering a new listener, CarDiagnosticService attempted to send
the last-known diagnostic frames to the listener as events. This turned
out to cause an issue with potentially dispatching null live frame data.

While this is easy to fix, the entire model of sending those events is not
especially useful for diagnostics, and stemmed from a similar concept in
CarSensorService, where it actually offers value.

Take this opportunity to remove the cached events entirely.

Bug: 37347349
Test: runtest -x packages/services/Car/tests/carservice_test/src/com/android/car/test/CarDiagnosticManagerTest.java
Change-Id: I6c8ed874e024400259bf84d8c71386878058b9d7