Explicit remapping of sensor IDs
OBD2 defines sensor indices in a way that is not identical to
what Vehicle HAL does. That is basically by design, as we group
sensors by their data type, and allow for vendor extensions.

obd2_app captures data using the standard OBD2 sensor indices,
but we want them to be sent over to the car/head unit in the Vehicle
HAL index mapping.

We were handling this in the injector directly, but as we add
more sources of diagnostic JSON that already use the HAL mapping,
the injector would need to know what sensor ID layout is in use.

Instead of adding complexity everywhere, just write a script
that can post-process obd2_app capture files and translate
sensor indices. From there onwards, sensor IDs are in the HAL
ordering, and can be sent via the injector as-is.

Test: manual
Change-Id: Ia36ca1c00658e139a9963b882a0bec7881f427fc