Add APIs to CarDiagnosticManager to check whether specific features are enabled
DiagnosticHalService has a DiagnosticCapabilities object that keeps track of whether
any/all the diagnostic HAL properties are supported by the underlying implementation.

This change exposes the same information up to CarDiagnosticManager for applications
to use, in order to quickly check whether or not they can expect support from a car.

Change-Id: I0fd05b1e24346483ea94432e428d0b0e56b7d7e4
Fixes: 38138457
Test: runtest -x packages/services/Car/tests/carservice_test/src/com/android/car/test/CarDiagnosticManagerTest.java -c com.android.car.test.CarDiagnosticManagerTest -m testIsSupportedApiCalls