Fix Bluetooth failing to connect after pairing
After successful pairing, supported profiles between 2 devices was not
connecting automatically, if the device was previously paired and
unpaired.  This is because, unpairing sets the priority of the profile
for that device to UNDEFINED.  Hence, subsequent pairing attempts fails
to connect, since the Bluetooth profiles ignore connection requests when
the priority is not set to ON or AUTO_CONNECT.
So, its upto the Device's Bluetooth policy to set the priority of
supported bluetooth profiles per remote device to ON, everytime it
pairs to a remote device. For Carkit, we do this in its
BluetoothDeviceConnectionPolicy. For phones, it is done in PhonePolicy.

Bug: b/38500587
Test: Tested by pairing and seeing the supported Car Bluetooth profiles
connect automatically. Also, checked the saved priority for the profile
per device in Settingsglobal content provider by querying
adb shell content query --uri content://settings/global

Change-Id: Ie102a0923022d844106bdb864cbd57c4ab5aa19c