Ajdust max/min height for notifications.
Reduce the max and min heights for notifications. This is causing an
issue where a media notification with an image would take up the height
of at least the min height. The value was too high.

Now, reduce the min height to be equivalent of a notification with a
title and one line of summary.

Test: booted up on AAE headunit. Issued several notifications, including
a media one.
Change-Id: I4ab5294d85ec2fe05e29ebbf4624a3b062b0d0ae