Replace test music with something easier to hear
Use a track with less dynamic range so that it is easy to detect volume
changes, etc.  The preivous classical track started with a significant
amount of silence, and frequent dramatic volume changes making it harder
to use for diagnostic purposes.
Track is from youtube music library and is noted as free for any use and
not requiring attribution.

Test:   Tap "Media Play" button on the Audio panel of kitchensink.
Change-Id: I0a41675c9d4dbfa81f6f35f5c746d2a4c5207853
Bug: 38389934