Flatten dependency hierarchy of legacy-android-test
Previous changes statically included legacy-android-test in preparation
for removing android.test.* and junit.* classes from the android.jar.
Unfortunately, that lead to duplicate classes between APKs and the
bootclasspath which caused build problems (Proguard) and also runtime
problems (when targeting and running on older releases).

Switching from statically including the classes to using the runtime
libraries cannot be done in one step because legacy-android-test is
statically included in libraries which are used in many APKs and so
removing it from those libraries requires that all APKs be updated at
once. Doing that atomically across dozens of projects is not practical.

This change modifies APKS that statically include the
legacy-android-test library indirectly.

* If the APK manifest uses the android.test.runner library then the APK
  is modified to stop statically including legacy-android-test and
  instead build against android.test.base/mock/runner libraries instead.

* Otherwise, the APK statically includes legacy-android-test.

Also, any libraries that statically include are modified to stop
statically including it and if it has source dependencies on the classes
is changed to build against the android.test.base/mock/runner libraries.

The following change descriptions were generated automatically and so
may be a little repetitive. They are provided to give the reviewer
enough information to check the comments match what has actually been
changed and check the reasoning behind the changes.

* car-support-lib/car-support.mk
    Added 'legacy-android-test' to LOCAL_STATIC_JAVA_LIBRARIES because
    EmbeddedKitchenSinkApp was statically linked to it through its
    dependency on vehicle-hal-support-lib

* service/Android.mk
    Added 'legacy-android-test' to LOCAL_STATIC_JAVA_LIBRARIES because
    CarService was statically linked to it through its dependency on
    vehicle-hal-support-lib

* tests/CarDiagnosticVerifier/Android.mk
    Added 'legacy-android-test' to LOCAL_STATIC_JAVA_LIBRARIES because
    CarDiagnosticVerifier was statically linked to it through its
    dependency on vehicle-hal-support-lib

* tests/obd2_app/Android.mk
    Added 'legacy-android-test' to LOCAL_STATIC_JAVA_LIBRARIES because
    Obd2App was statically linked to it through its dependency on
    com.android.car.obd2 which in turn depends on
    vehicle-hal-support-lib

* tests/vehiclehal_test/Android.mk
    Added 'android.test.base' to LOCAL_JAVA_LIBRARIES because
    VehicleHALTest's source depends on its classes and because of these
    changes they are no longer present on the compilation path. The
    classes do not need to be statically included because the classes
    will be provided by the runtime, either from the default
    bootclasspath or from the android.test.runner library that
    VehicleHALTest specifies in its manifest.

* vehicle-hal-support-lib/Android.mk
    Removed legacy-android-test from LOCAL_STATIC_JAVA_LIBRARIES
    because vehicle-hal-support-lib is not a package so does not need
    to statically include the classes

Bug: 30188076
Test: make checkbuild
Change-Id: I46b28d0092d561b3f047480f1ea8f330ed1d8535