Revise CarVolumeGroup to syncrhonize gains
Move CarVolumeGroup to synchronize the gains of its constituent buses in
millibels rather than arbitrary index space.  This allows for buses with
different min/max values, for example.
For now the mapping is strictly linear, but this code structure is
intended to allow for alternative group index to gain response curves in
the future.

Bug: 73791816
Test:  TBD
Change-Id: I589d806bbeaec4bf679d7f5b5ae3b9e79ac34ea9