Add car specific focus policy handler
The car specific focus policy installation accomplishes two things:
1) By giving car code control over which audio focus requests get
granted, it allows cars to concurrent audio focus for different types of
audio.  We'll see if this ends up being a good thing or not once we
start experimenting in actual use cases.
2) By having the a focus policy installed, we bypass the built in
ducking logic, leaving that to the car to handle below the HAL.

Test:  build and run KichenSink on Mojave and listen to radio in Jeep
Bug:  33352354
Change-Id: If2dfce9ad4620a4d840b95f9a2e2f67733f3b155