Construct default audio port config when creating audio patch
It's not trivial to get activeConfig for a device port unless we copy
the activeConfig from the mix port used to route to this particular
device. There can be more than one mix ports routed to the same device.

App is also given the privilege as to select the device port
configuration therefore there is guarantee the activeConfig from the mix
port is used for that device port.

In this change, we construct the source and sink config for an audio
patch with default values and it's up the audio HAL to pick the right
configuration.

We will update the go/aae-audio doc accordingly.

Other changes in this CL:

- Changes the getExternalSources as to return list of addresses
- CarAudioPatchHandle uses address instead of name
- Also does code cleanup.
- make update-car-api updates proguard-release.flags

Bug: 77147937
Test: releaseAudioPatch(createAudioPatch(<test-port>, <usage>, 0))
Change-Id: Ia30c2bfde8198546bfa9090882d424ca0ba14363