Enable dumpsys car_service
The command in use to dumpsys information from CarService was "dumpsys activity service com.android.car/.CarService", which is based
on overloading dump() in the CarService class. OTOH, the class known to the ServiceManager as "car_service" is ICarImpl, which was not
providing the correct overload of dump().

Fix this by moving the dumping logic over in ICarImpl

Test: manual
Fixes: 77498074

Change-Id: I0c18a9088e9c06201de79e8ebf247dc969ff7317