Deprecate the LocalBroadcastManager
LocalBroadcastManager works within one process and since the car trust
agent runs in user-0 as a singleton, secondary users won't be able to
use LocalBroadcastManager to communicate with the underlying service.

This also reduces the final apk size due to the removal of support
library.

Due to the asynchronous nature of TrustAgentService in framework,
for intance, the result of addEscrowToken would be delivered in a
separate onEscrowTokenAdded callback.

When enrolment client tries to add escrow token, followings would happen

- Enrolment client calls
CarTrustAgentBleService.setTokenResponseCallback and sets itself as the
callback
- Enrolment client calls CarTrustAgentBleService.addEscrowToken
- CarTrustAgentBleService delegates the call to TrustAgentService
- onEscrowTokenAdded is delivered to TrustAgentService and callbacks to
enrolment client

What's next:
- CarEnrolmentActivity should not store the handle information,
  it should be CarTrustAgentBleService's responsibility
- Associate the token/handle with user id, see b/77854782

Bug: 76008345
Test: manual
Change-Id: I9f28f14cd7ae08786ef3b2825735b15c7de80507