Scrubbing the list of whitelisted apps w.r.to UX restrictions.
Removing the internal apps from the list of exempted apps that are
not blocked in a UX restricted state.  The whitelist will be overlaid
from a vendor overlay as needed by the OEMs.

The ones that are kept here are to notify which system apps are expected
to be whitelisted by the current implementation.

Bug: 73537474
Test: Vendor overlay works.

Change-Id: I6335c19b8f58465dbd7a28f483dd40db34d2738f