Fix ManifestMerger issues
Set LOCAL_MIN_SDK_VERSION in CarService so that the build system
doesn't raise the minSdkVersion value specified in the manifest, which
causes ManifestMerger to error when trying to merge it into a package
that specifies a lower minSdkVersion.

Point the android:label from CarService to an app_title resource
string, which matches the tests that embed CarService using
car-service-for-tests and prevents ManifestMerger from erroring on
merging manifests with different values for the application label.

Bug: 78447299
Test: m java
Change-Id: I8c4b2119419d745d0b39f2793153417c8f2efce6