Convert android.car to Soong.
In order to mirgrate Car java docs to Soong, we have to convert android.car
first. Synced with people from Car team, the original owner for the
flag:(TARGET_USES_CAR_FUTURE_FEATURES) have changed team. It seems that
currently this flag was set to true by default. So removed this flag,
and always use the future sources as Android.bp doesn't support
if-condition.

Submit this CL first before we make any changes on droiddoc in case we
break too much post-submit targets.

Test: lunch bat_land-userdebug && m -j android.car
Bug: b/70351683
Change-Id: I1135a84b1fd0e684c85a491c3811dd7e62bae721