Listen for CMPS Shutdown prepare and disable bluetooth adapter
Bug: b/117243081
Test: Manual. Power the car on, pair a device, turn the car off and open
a door (this gets SHUTDOWN_PREPARE to fire from CPMS). Verify that the
device has disconnected from the car and that the car's Bluetooth
adapter is off (dumpsys bluetooth_manager). Wait for Android to fully
shut down (your adb session will be forcefully ended) and power the car
back on. Make sure Bluetooth starts back up. Additionally, make sure
when Bluetooth is turned off before SHUTDOWN PREPARE is fired that the
off state is persisted on reboot and this module doesn't try and enable
Bluetooth.

Change-Id: Iac5d0e9c09c2b67646c85d0008df5853a6cdc4fc