car.usb.handler
Added some logic to the car.usb.handler to allow it to resolve a USB
handler as soon as the device running AAE boots in the case where a
USB device is connected before AAE is started. This logic already
existed to some extent, but now it is looking for the locked boot
complete intent which allows for the USB device to be resolved sooner.

Change-Id: Icb1148ce0f4e5609a9f9b57e62db0d9809943995