CarProjectionManager: Introduce ProjectionKeyEventHandler
This API will be used in the place of CarProjectionListener to allow
projection clients to register to handle specific input events from the
system. Unlike the previous CarProjectionListener API, this API can be
extended to additional input events as needed in the future.

This new API also supports the following new events:
- Immediate notification of key-down for the VOICE_ASSIST key
- Key-up after a long-press of the VOICE_ASSIST key
- Short-press, long-press, key-down, and long-press-key-up events
  for the CALL key

Test: Manual test, atest CarServiceUnitTest CarLibTests
Bug: 129706517
Change-Id: I93a2c1551738e44e9240b2a42f04a804e0d13543