Return and use connect/disconnect return codes + ignore PRIORITY_OFF
It was found that the returned priority values default to PRIORITY_OFF
when the profile proxy has an error. This was causing devices to get
removed from the reconnect list when they shouldn't have been. Instead,
we plan on leaving devices in the list and allowing connection attempts
to immediately return when the proxy determines they've failed (returns
false). It's assumed the profiles will NEVER connect a PRIORITY_OFF
device and immediately return false.

Note that this required updates to the tests as well.

Bug: b/132439469
Test: atest CarServiceUnitTest
Change-Id: Ib48abbdfae8ede09e523a47f9a4183057ba555cb