Added Audio Focus test for Car Audio
The requirements for audio focus are slightly different for cars, the main
difference stems from the model the OEM requires in the car. In particular
OEM would like to concurrently output different audio usages,
typically each audio is routed to a different device.
The test written separates the three interactions rules for audio usages
when requesting focus:
Exclusive Interaction:
Focus is exclusive and focus is always requested exclusively for
each usage. In the interaction between requestor and focus holder, a
focus lost will sent, the focus may be transient or permanent
dependig on the focus requested and the current focus holder ducking
options.
Reject Interaction:
Where the requested focus in this interaction is always
rejected regarless of the focus requested.
Concurrent Interaction:
In the interaction between the current focus holder and focus requestor
the two can hold focus concurrently if the focus requested is of type
AUDIOFOCUS_GAIN_TRANSIENT_CAN_DUCK. In this case the focus loss is not
dispatched to the current focus holder as no focus change occurred.

Bug: 130257985
Test: atest com.android.car.audio.AudioFocusTest
Change-Id: Id76c4e913f8d5601e0855ad8228fe6726a34f636