Expose TrustedCredentialsSettingsActivity in CarDeveloperOptions.
This activity is used by the CtsVerifier app for a step in CAInstallNotificationVerifierActivity. Since advanced Wi-Fi features were descoped for CarSettings, this test case is an automatic fail without UI to verify a user certificate.  Developer options must be turned on during CTS, so this will provide the UI necessary to pass the test without being a user-facing feature.

Bug: 130199964

Test: build and deploy, enable dev options, install and run CtsVerifier, ensure test case launches UI.
Change-Id: Ieba8f6b5052ee9477bdad575132e64fe941bb27f