Refactor OperationType into its own proto.
This allows the OperationType to be used within public facing APIs
without being tied to a specific BLEMessage version.

As part of this change, rename the proto files so that they follow best
practices for naming:
https://developers.google.com/protocol-buffers/docs/style#message-and-field-names

Lastly, align the proto package name with the package name of the iOS
protos. The java specific class names have also been renamed to have the
same outer package.

Bug: 133155675
Test: Test enrollment and unlock flow with iOS.
Change-Id: I4d5cd8af70f867d59dc6f12f5f4e57646a15ceaf