Add a dummy launcher for CTS
Some CTS tests (especially MultiDisplayClientTests) are failed in AAE,
since the CarLauncher creates ActivityView and launch Map Activity on
it, and whenever Home is requested, it'll try to restart Map Activity on
the ActivityView.  This makes unexpeceted events in the view of CTS and
the tests will fail.  So AAE team tries to pass CTS by replacing CarLauncher
with this one.

Bug: 133274622
Test: atest CtsWindowManagerDeviceTestCases:MultiDisplayClientTests
Change-Id: I6c0b2fef98db1fa9603de71fec9d9169736873f8