Set the CarBleTrustAgent to be the default Trust Agent.
Enable only the CarBleTrustAgent in the automotive  builds, in
case there ends up being multiple agents in the system.  This
doesn't quite fix the problem with Kitchensink also showing up as a
Trust Agent, but it mitigates that issue by enabling only the
CarBleTrustAgent.

Also, removed revoking Trust during onDeviceUnlocked, since that is not
really needed to ensure user switching does not automatically
authenticate the user.

Bug: 133792836
Test: Manual
Change-Id: Id95e96a7a9c395f2445e44a11925194ad8c05f28