Add progressbar to BugReportApp.
When BugReport is initiated when dumpstate is in progress
by the BugReport app, it will show a dialog with a progress
bar and numeric value of the progress.

When dumpstate finishes, it changes the title to "bug report collected",
it doesn't close the dialog.

Bug: 134530860
Change-Id: I2e768d47b8122e3cd67e7872c5373d901cc8e3d6
Test: tested on a hawk rig - took bugreport, uploaded,
      clicked start bugreport shortcut when it's running,
      it shows a progress.
(cherry picked from commit ca85bfd52e652589f1fbf7dfecc50a6c003a0242)