Add screencapture feature to bugreporting
Dumpstate does not collect screenshots of the physical displays other
than the primary one. Add logic to collect these because it is important
for Automotive use case.
Further, Android bugreports does not include screenshot of the primary
display in the bugreport (its api allows capturing it seperately). The
API in automotive combines all displays in a zip file for automotive.
Bug: 133368541
Test: Manual
Change-Id: I412fb55e2d0bd734b264014857ad80e4ec04870e