Break down the evs app
Make a copy of the original evs_app code and break it down into two
parts: most of the logic goes to support lib, and the rest including
main function goes to a new app.

Make minimum changes to get the code workable. There will be follow-up
changes to create use case in support lib.

Bug: 130246434

Test: build and run on Hawk device

Change-Id: I0ece1f804fcdf1c40e37e6e03ae31a3083017e81