Fix CarUsbHandler to resolve all AOAP devices immediately.
The expected behavior of connecting an AOAP device is
that if there is exactly one AOAP handler installed on embedded, the
AOAP device should automatically be resolved with that handler
(otherwise, the handler picker dialog should open). The devices that
have more than one possible handler do not automatically run the aoap
handler because the device is not in aoap mode.

This change modifies the UsbHostController logic to just look to see if
the device supports AOAP, and not whether or not the device is already
in AOAP mode. The UsbDeviceHandlerResolver already requests an AOAP
switch if the handler is an AOAP handler, so there's no need to be in
AOAP mode before we attempt to resolve the handler.

Bug: 135569578

Test: Connect an Android MD that has multiple possible usb handlers
(Samsung MDs exhibit this behavior) which has the Android Auto app
to a Hawk bench. Android Auto Projected should start via aareceiver
without any handler selection required.

Change-Id: Idc482edbacbc7a8ae2719ae36995dfeaad1b71bc