Fix and improve the test script
The test script was missing to copy validated files (actually it was
doing it only for screenshot) due to a bug in logic operations. Also
rather than creating the zip file in a separate directory, create in a
unique temp directory in current dir, so user have a way to further
inspect the zip file if necessary.

Test: Manual, by running the script
Bug: 135715013
Change-Id: I3d8a3d4cb68a1e3ae3cf4ef5f54c6bc327a263a6
(cherry picked from commit 07d91f99be4110a4ce09658593064d6949dfa844)