Added logic to select route sound to left speaker
When emulating multizone audio, the left speaker can be used to emulate
a zone. The property aae.playZoneId will be used to select which zone is
actually played in the right speaker. Logic was added to kitchen sink so
that when the zone is switch the correct zone is set by using the
playZoneId property.

Bug: 134524865
Test: Ran zone switching in emulator.
Change-Id: Ie1d2c5ef51333cc8544aa27aa7621fd15d3b792d