Implement save-to-USB for bug reports.
In this new functionality, a user has an option to either upload the
report, or move a report to external storage of their choice. Once
the user pushes "Move" button, a prompt asks user to select external
directory where to save the bug report to. By default, the reports are
still uploaded automatically.

Bug: 135555542
Test: manual testing
Change-Id: Ibd25f939f460c375c2a16933fba2f85b81a1f93b