Remove abuse of setupBluetoothConnectionProxies to wait for proxys
CarBluetoothUserService maintains a set of Bluetooth profile proxies to
interact with the Bluetooth stack. These proxies use a service
connection that let our class know when its okay to use them and when it
might not be. The async pattern posed problems for the one-shot device
connection code that would timeout on connection attempts if the proxies
weren't ready. CarBluetoothUserService remedied that by blocking until
the proxies were connected on setup. Setup was abused in other functions
to block if a profile wasn't connected though, which would create
proxies unnecessarily and recursively wait forever if proxies didn't
connect. This caused lots of broadcast-intents to be sent to us, as well
as potential stack overflow exceptions.

This change simplifies the work done on setup and pulls the code that
blocks until all proxies are up into a separate function that others can
use. The recursive profile connection check is remove entirely in favor
of using a single waitForProxies() call if its found your proxy isn't
connected.

Bug: b/137692872
Bug: b/137753671
Test: atest CarServiceUnitTest and automated toggle testing that checks
that devices are connected on the correct profiles each time after
Bluetooth starts without any "Proxy Unavailable" errors.

Change-Id: Iefa3ace96995843839a47bec1f180daa8f8b6f0d