Remove obsolete broadcast behavior
There is no need to broadcast when a footer is displayed/removed. The
broadcasts were added specifically for gmscore to know when the footer
was visible, so that GLS consent could be given. However, this
was only very briefly actually used within GmsCore, and has been
obsolete for several releases.

Test: CTS
Change-Id: Ie9775b1a8c66372ea128fbb8bdf69249f7ce9b3f