[AAE Garage Mode] Change the UI layout of the Garage Mode Test App
On narrow displays, one pulldown menu is so narrow that its setting
is not visible.

Rearrange the UI elements to put each pulldown menu on a line by itself.
This way each has enough space to show its settings.

Also initialize the "Require idleness?" button ON. Idleness ON is
required for jobs that run during Garage Mode.

Fixes: 138670679
Test: Verified on Hawk
Change-Id: I81f8530f2c0857662a5d89aaec5e2fbec36ec074