Changed android.car.testapi build back to using soong
The build file for building android.car.testapi was changed from using
soong to make. This was because testapi required a dependency that was
built with make. This dependency has since been converted to use soong,
so the testapi can use soong to build as well. There was also a
regression introduced in the original conversion from soong to make,
which caused it to not be copied to dist in dist builds, and this change
addresses that as well.

Test: Verify android.car.testapi.jar is in dist artifacts for the
hawk-userdebug target on master.

Change-Id: I1bb9533585230d3251a91823c95cf8ac49acc742