Change the launchMode of the UsbHostManagementActivity to standard
The launchMode of the UsbHostManagementActivity is currently set to
singleTop. This causes issues when there is more than one USB device
connected, and they are not all aoap devices. Each USB device should be
processed in its own activity, but since the activity is being launched
singleTop, there is only one activity instance, and it just receiving a
new Intent for each USB device, instead of starting a brand new
activity. Changing the Activity to standard allows a new Activity to
start for each device.

Bug: 138004874

Test: Flash Hawk in car. Hawk in car has at least one other USB device
(USB to Ethernet adapter). Boot Hawk (start car)  and run embedded first
run setup. Modify the setting in the Embedded Settings App under More->System->Android Auto called
"Launch Android Auto on Phone USB Connection" to be enabled. After finishing setup,
shutdown the Hawk, and plug in a phone which supports the Android Auto App
(and has it installed). Boot the Hawk, and verify that Android Auto
projected starts on boot.

Change-Id: Ie26a4a130aca5071c2f28e18863c61174d06f8b9