Reland: Rework platform version to hide codenames.
The public platform version no longer can be a codename, it is
always the most recently released platform.  A new build property
and API provides either the offical version or the current codename
as appropriate.  This will avoid breaking apps that look at the
platform version while development is under a codename.

Bug:  143175463
Test: manual
Change-Id: I77b29ece0c4f39d444811ae5e4eaa163e9d66e12
Exempt-From-Owner-Approval: Previously approved, no deltas in this project.