Move the app to /system_ext
To unbundle product partition, we plan to forbid hidden API in product partition.
So this app should be moved to /system_ext because it uses hidden API.
It is not permanent, so they can be moved back to /product due to
needs if there isn't hidden usage anymore.

Bug: 137908189
Test: m
Test: check whether device boots successfully
Exempt-From-Owner-Approval: cherry-pick from internal
Change-Id: I9c4fc342a180aa938cde839cff967f58cfc38df3
Merged-In: I9c4fc342a180aa938cde839cff967f58cfc38df3