Fix crash in CarPowerManagementService
ag/9718627 moved the logic for when we call getInitialUser().  This was
causing us to try to get the initial user before the first user was
created on initial boot.

This CL updates the logic to only try to get the initial user when
running the resume flow.

Bug: 144802294
Test: Flash and check logcat for crash.  100% repro before change, 0%
repro after change.
Change-Id: If500c1ccdb9655c46edb1745b007186757db0f2b
(cherry picked from commit 0e6e81fc86fdf340503d3e416ad274ae29817d37)