Changed STR resume behavior when current user is an ephemeral guest.
Previously, when the system resumed from suspend and the previous user was an
ephemeral guest, the system would switch to the first non-persistent user (which
is not a guest). The main reason for this restriction is privacy - the vehicle
could have been suspended for hours and the (human) user "resuming" it might not
be the same person that "suspended" it.

This CL changes that behavior - while preserving the privacy - by automatically
creating (and switching to) a new guest user in this scenario

Here's how to test it:

1.Create a guest user (say, 11)
2.Switch to that user (adb shell am switch-user 11)
3.Suspend (adb shell dumpsys car_service suspend)
4.Wait until garage mode is done (watch for logcat CAR.POWER)
5.Resume (adb shell dumpsys car_service resume)
6.Notice a new guest was created and switched to

Test: manual verification (see above); new unit tests will be added later.
Test: atest CarServiceUnitTest:com.android.car.CarPowerManagementServiceTest

Bug: 146020398
Bug: 146380030
Bug: 145027829

Change-Id: Icb486877385985cdaab50914e3bbf4371685f410
(cherry picked from commit 80a47e10307e73c4dc02d59d1ccb453944f08965)
(cherry picked from commit 15dd57ade16be5f5c0d6ac1fb193b421db9960a9)