Build EVS reference driver as a vendor process
With this change, the reference EVS driver will be built as the vendor
process.  And, to achieve this, all libgui usages are replaced with
newly introduced Automotive Display Service.

Configuration XML file is relocated from /system/etc to /vendor/etc.

Also, this change deprecates old instance names for EVS Enumerator HAL
implementations.  Instead, we will use regex name, [a-z]/[0-9],for HAL
instances and new method, isHardware(), to tell whether it is EVS
manager or not.

Bug: 147010694
Test: VtsHalEvsV1_1TargetTest
Change-Id: I32547861018389c86abdfe16a608b7722d3dc8b8
Signed-off-by: Changyeon Jo <changyeon@google.com>