MA-16248 fix CTS neverallow rule check failures on evs HAL
evs HAL is moved from system partition to vendor partition, while binder
is still necessary, so the system neverallow rule was modified so that
domains associated with "evs_domain" can use binder.

associate "hal_evs_driver" with "binder_in_vendor_violators", this is a
exception in the original system sepolicy, and then the related system
neverallow rules are not violated.

Change-Id: I33ae96e890bf1ff54aabc6605a9547f21c99043e
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>