MA-16255 The parking line in AP drawing reverse
the parking line between m4 and ap is not align, the direction is contrary.

First check whether the M4 line or the AP line is wrong.
It was finally confirmed that the AP's line was wrong.
Then find the line drawing code.
Read the line drawing part of the code and change the reverse part.

Change-Id: If94581074644ed11da00a0dfd24c822a519b8d2f
Signed-off-by: zhipeng.wang <zhipeng.wang_1@nxp.com>