Fixes KitchenSink flickering issue.
Flickering is caused by texts trying to be in the
center, though the texts length keeps changing. Fixes the issue by making texts
aligned on left side of the screen and padding some space on the left, so
that users have a good visibility of texts on the left side near screen edge.

Bug: b/144698249

Test: Running the app on a device.

Change-Id: Iab82c724def425d12e9ad19d8f0675365723364c