Drop useless check for PRODUCT_BOOT_JARS redefinition.
It is not clear why this check exists; it seems to be unneeded, and
removing it does not break any tests.

PRODUCT_BOOT_JARS is a product variable, and therefore handled in a
special way (unlike ordinary Make variables). It is substituted with
a macro that inherits makefile-specific instances of this variable
and merges them into one. This can be seen by adding debug output
to the device-specific makefile.

For example, add the following to device/generic/car/aosp_car_arm.mk:

  $(warning $(PRODUCT_BOOT_JARS))

The build will generate a warning similar to this:

  device/generic/car/aosp_car_arm.mk:21: warning: @inherit:build/make/target/product/aosp_arm.mk @inherit:device/generic/car/emulator/aosp_car_emulator.mk

Test: lunch aosp_car_arm && m

Change-Id: Ib5e09c6624dbd4db29105c67543a53caa490b818
Merged-In: I061b70b751ca00efa8f215dba8419d2b9e920403