[Unit test] Shorten CarPowerManagementTest execution time
IRL we poll every 2 seconds for changes in the power state. To
speed up the testing, do this every millisecond. Also, some
tests wait to verify that a transition did NOT occur. Reduce the
time that we wait for that.

This CL also fixes two issues that contributed to flakiness:
 - Ensures consistency in CarPowerManagementTest between the list
   of received power states and the semaphore that indicates that
   a new state has been received
 - Adds a null check to CarPowerManager

Fixes: 130378693
Test: This updates CarPowerManagementTest
Change-Id: I54a07dcf852707247c6bf2ba81b2c3418cf5815a