Fix for stride length coming out from evs manager.
InputFrame data structure expects stride to be in bytes where as
hardware buffers on android output stride in number of pixels. This is
different from the format used by mediapipe graphs where stride is the
number of bytes per row. As number of bytes per row is the more general
convention, we continue to use that for input frame while converting the
stride parameter received from evs to reflect this.

Change-Id: I6dde4664603d906d5c4c58c8e3435a472033f7f2
Fix:149030425
Test: m -j