Clean up API state of CarSettings.
This class only provide constants, which only one of them is a @SystemApi -
everything else should be gone.

In theory we would need to mark those as @removed, but it's very unlikely
that anybody is using them (and the automotive ecosystem is still small
enough to let us make these breaking changes).

Bug: 150169480
Test: m -j android.car-test-stubs-docs-update-current-api android.car-stubs-docs-update-current-api android.car-system-stubs-docs-update-current-api

Change-Id: I3757608110eac2e3bb076558e87b7d244e2ab124