Fix multi-thread issues in StreamHandler
In current design, it is possible that the client stops a video stream
before the last analyze thread is scheduled.  Because stopping a video
stream sets an analyzer callback as null, the analyze thread will
therefore be crashed as soon as it is scheduled.

To prevent this, this change adds a mutex to protect shared variables
and adds a nullcheck in a analyze thread.

Fix: 150255890
Test: m -j libevssupport and run face enrollment w/ user switching
Change-Id: I161be8d69f8116846f3863a6e60e491ccf1af2ca