Present user-facing UI when VHAL fragment cannot be used
With SELinux enforcement turned on, the Vehicle HAL fragment does not
have permission to connect to the HAL instance.

This could be fixed but it would require a decent amount of effort,
and in general the VHAL tool should only be really useful during early
bringup, when SELinux is likely to be turned off anyway. In all other
scenarios, the car property test UI should be preferred.

Bug: 150336686
Test: setenforce 1 and load VHAL fragment
Change-Id: I4070cf786d27bde0896a6e5576f0bd3aab5a728a