Update the current API dump
All aidl_interface modules should by default considered as stable, in
case it is used across system and vendor partitions, or across modules.
Like other API surfaces, we need to have a dump for the current
(yet-to-be-released) version and update it when there is an API change.
This is done via m <name>-update-api.

Then the owner of the interface can freeze the current version as a
numbered version via m <name>-freeze-api.

This change shall be rejected only when the owner is certain that the
interface is not used across the updatable boundaries.

Bug: 152655547
Test: m
Change-Id: Ibf2f38482915bc49132c378bc01b76b369a1a17d