Allow power transition from "Wait for Finish" to "On"
Some OEMs want to bring the system to the point of shutting down,
but hold it there for a few minutes. They may then either shut
down or bring the system back up.

Previously, once all the preparations for shutting down were complete,
the only permitted power transition was to shut down. This CL allows
the shut down operation to be canceled even when everything is all set
to power down.

Fixes: 153755740
Test: Created CarPowerManagementTest.testCancelShutdownFromWaitForFinish
Change-Id: I57745a9c8267968c48db8fe7de083c5d4af27341