De-flake and clean up CarPowerManagementServiceTest
Two of the tests are flakey, this seems related to test setup
behavior specific to these two tests. After recent refactoring,
it is not needed to set a power-state listener. Removing that
eliminates the uniqueness of the flakey tests.

This CL also removes some unnecessary method parameters and
clarifies some logic.

Fixes: 154927360
Test: 'atest CarPowerManagementServiceTest' on emulator
Change-Id: I85a84e4714ec0c294456d49c61da5a3a4d565977