Fixed delayed audio focus abandon request.
Fixed delayed audio focus to perform correct abandon logic for
AudioFocus info that do not pass in the audio focus gain info. Added
test to CarAudioFocus to test for full integration test. Note that this
cases are already present on CarAudioFocusUnitTest but failed to be
detected due to AudioManager logic.

Test: atest CarAudioFocusTest
Test: atest AudioFocusTest
Bug: 156114894
Change-Id: Ia301f9761c9c23d33aebae5442e1033df27e7e6e