Add Silent Mode for Automotive
This monitors the Silent Mode indication from the kernel via
/sys/power/kernelsilentmode.

The indication is available from SilentModeController to other code via
a callback or by direct query.

This new indication is used to control turning the display on at
system start up and off at shut down.

New reboot options for test are included:
   adb reboot forcedsilent
   adb reboot forcednonsilent

Support for
   adb shell cmd car_service silent <on|off|query>
is in a dependent CL.

Bug: 134521909
Test: atest SilentModeControllerUnitTest (new)
Test: atest BluetoothDeviceConnectionPolicyTest
Test: atest CarBluetoothServiceTest
Test: atest CarCabinManagerTest
Test: atest CarPowerManagementServiceTest
Test: atest CarPowerManagerUnitTest
Test: atest CarUserNoticeServiceTest
Change-Id: I4ec16ec2a7f5504de1184f982d31d1c83ffcf469