Modifies HalCamera::clientStreamEnding() method
This change modifies a declaration of HalCamera::clientStreamEnding()
method to take a raw pointer of VirtualCamera object instead of its
smart pointer object.  As this method is called from VirtualCamera
destructor, it causes a fatal error when a passed smart pointer object
is being destroyed at the end of the method.

Also, this change moves a line to clear a list of HalCamera objects in
VirtualCamera::shutdown(), to ensure it does not lose them during the
resource clean up.

Fix: 158339680
Test: Run /system/bin/evs_app in test mode w/ a logical camera device
      and check there is no signal when it closes the device.
Change-Id: I0897bd7b55982f34162304d305e30666f3cb8b2f