Fix map of virtual displays to physical ports
mPortLookup is map of display ids to physical ports. For virtual
displays we had incorrectly saved a map of the virtual display's
id to its physical display id.

This wasn't caught earlier because most of our test devices seem
to use the same id for the display and physical port.

Bug: 154263657
Test: atest CarUxRestrictionsManagerServiceTest
Change-Id: I1ba01c3277fcb2406c96dcef76ae7ec0d80b8059