Add emulated camera device support
This change implements logics to emulate camera devices for EVS clients.
As the first milestone, this synthesizes a video stream from images
stored on the device in YUYV format.  Source images must be in the same
resolution and accessible by automotive_evs user/group.

This feature helps the developers to verify and validate their EVS
service implementations on predefined static image sequence and
therefore also debug the issues more efficiently.

"(?i)emulated/[0-9]+" pattern is reserved only for emulated camera
devices.

Below example command adds an emulated camera device as "emulated/0" and
configures it to generate 640x360 video stream from images stored at
/data/vendor/evs/, at every 100ms.
$ adb shell lshal debug \
  android.hardware.automotive.evs@1.1::IEvsEnumerator/default \
  --configure-emulated-camera emulated/0 /data/vendor/evs/ 640 360 100

If a emulated camera device with a given identifier is active, updated
configurations will be effective when it is closed.

Bug: 157678559
Test: Run above example
Change-Id: Ib2610f9d2f5bd07246abbf36be07eeaee9ef59d1