Support PNG format image files
This change implements logics to synthesize the video stream from a
sequence of PNG format image files for EVS Camera HAL emulation.

If the source image is larger than the output stream size, the user
configured, the image would be cropped in the output stream size at the
left-top corner.  If the source image is smaller than the output stream
size, the frame buffer would be filled partially and the source image
would be located at the left-top corner.

To use a similar logic to read, this change also modifies a frame dump
logic in the reference EVS HAL implementation, to write additional
information such as the width and the height.

Below example command adds an emulated camera device as "emulated/0" and
configures it to generate 1280x720 video stream from images stored at
/data/vendor/evs/, at every 100ms.
$ adb shell lshal debug \
  android.hardware.automotive.evs@1.1::IEvsEnumerator/default \
  --configure-emulated-camera emulated/0 /data/vendor/evs/ 1280 720 100

After updating config.json file to use emulated/0 for the rear view,
below command starts showing the preview from the synthesized video
stream:
$ adb shell evs_app --test --gear reverse --display 2

Bug: 162946784
Test: Run above example
Change-Id: I7218339d6e6369bcec1ba2b1adcbe8230a1fd0b8