Added key setting for car volume key event timeout.
Added setting for controlling timeout during which a car volume group
will be considered active (after playback has ceased in said volume) for
the purpose of volume control during volume key events. If the
volume group (or any of car audio context in the group) were active
prior to this timeout, a volume key event will trigger a volume change
for the volume group depending on the highest priority context that
can still be consired.

Bug: 173141906
Test: build and run
Change-Id: I2dbdf2e806915998f81bca3d7f7981fc1b055373