Disable verify_uses_libraries check for test apps.
Disable verify_uses_libraries check for test apps by setting
"enforce_uses_libs" option to false in the Android.bp files. This is
done automatically for android_test, but for tests defined as
android_app the check has to be disabled manually for each case, as
such tests are indistingushable from usual apps.

This will prevent build failures when the default for
"enforce_uses_libs" is globally switched to true.

Likewise, dexpreopt has to be disabled manually for these tests.

Bug: 132357300
Test: m nothing
Change-Id: Ib1b28f7940f4e75e90a1e7555d08a50874075f36