Added timeout for already selected context.
If a user begins to change volume via a volume key event (or volume
knob) on a particular context, the volume key event will be applied to
the context during the timeout period (even after the context has become
inactive). This allows the user to continue to change the volume on
transient sounds.

Bug: 173141906
Test: atest CarVolumeTest
Change-Id: I56b428bfadc23937ec4a0dfbcb7d96780f99ff7f