Fixes bugs in EVS manager
This change fixes two below issues in EVS manager implementation.

1) accidental frame request loss
When the client holds the maximum number of frame buffers it is allowed,
the client does not accept any new frame buffers and notifies EVS.  EVS
manager, however, regards corresponding frame request as done regardless
of what the client returns and therefore makes the client waits for a
new frame until a request is timed out.  To address this issue, this
change makes the client also marks its frame request is done so files
another request if it is still running.

2) slowness of the resource clean up procedure
When the client requests to stop a video stream while there is any
pending new frame requests, it will be blocked until a pending request
is timed out.  Therefore, this change modifies a predicate to stop
waiting for a new frame when a state is not RUNNING and awakes the
capture thread before it tries to join it.

These fixes are verified by running EVS reference application and then
launching EvsCameraPreviewApp.
> adb shell evs_app --test --gear reverse --display 0
> adb shell am start -n \
  com.google.android.car.evs\
  /com.google.android.car.evs.CarEvsCameraPreviewActivity

Fix: 177062717
Fix: 177223398
Test: Run the scenario above
Change-Id: If99e3beb4c0f9e85c82871bc3337a35d2372e6ab