FastPairProvider
Update Car Service to advertise a Fast Pair Provider anytime the vehicle
is in Bluetooth pairing mode.  This service makes it easier for supported
phones to detect and pair with the vehicle in addition to downloading a
companion app if necessary.  This service can be enabled by defining a
Fast Pair model id in the corresponding device overlay.

Bug: 171817451
Test: atest BluetoothFastPairTest
Test: Fast Pair Validator
Change-Id: I69f51eb58b3461a5258cbdb7264697fb8624a8aa