Add FakeCarAudioService to allow FakeCar usage in google3
Replaced the usage of SparseArray.contains method with
SparseArray.indexOfKey. Because contains is introduced in API 30 it was
crashing when called on older platforms.

Replaced @Mock ICarAudio with a fake implementation of the service.
Because ICarAudio.aidl depends on android.media.AudioDeviceAttributes,
and AudioDeviceAttributes is @hide and is living in the framework, @Mock
will not work anymore.

Bug: 181372746
Test: atest CarLibTests
Change-Id: Ica005716b28280372095503e93df68c1972b79a2