Add BluetoothConnectPermissionChecker to Kitchen Sink app.
Behavior now is to show a dialog asking for BLUETOOTH_CONNECT permission
when the BT Headset or BT Messaging menu items are selected. If the
permission is not granted for BT Headset, it will show a toast and
disable the Pick Device button. If the permission is not granted for BT
Messaging it will show a toast, and NotificationListener will not be
registered to receive device updates.

Test: manually confirmed behavior described above
Test: opened both menu items with permission not granted, then granted
Fixes: 187232849
Change-Id: Iee8663e105a421150d0fc83aaad03c5ed5e02bd1