Modify HalCamera::unsetMaster() method
When VirtualCamera object is being destroyed, it calls
HalCamera::unsetMaster() method with a 'this' pointer.
HalCamera::unsetMaster() is declared to take a strong pointer object of
VirtualCamera so it instantiates a temporary strong pointer object and
then destroys when it returns to the caller.  As there is no reference
on this temporary object, it eventually calls the destructor of
VirtualCamera object that is already being destroyed.

VirtualCamera::shutdown() checks its current state and returns
immediately if it is not in RUNNING state.  And, therefore, the system
destroys VirtualCamera object before it completes the shutdown
procedure.

To fix this issue, this CL modifies HalCamera::unsetMaster() to take a
raw pointer to VirtualCamera object, instead of a strong pointer.

Fix: 161762538
Test: m -j
Change-Id: I3ab7cbe289129a6c43357f03a4fe2167ad0d2ba5