Reformat repeating threshold tags in XML configs.
Reformat repeating thresholds such that they are enclosed within an outer
tag.

System-wide thresholds are enclosed with systemWideThresholds tag
and individual thresholds are enclosed with in alertThreshold tag.

Package specific and app category thresholds are enclosed with in
packageSpecificThresholds and appCategorySpecificThresholds tags,
respectively and individual thresholds are enclosed with in
perStateThreshold tag.

Test: atest libwatchdog_test
Bug: 186445673

Change-Id: I1ede554bc5d0b1478c669f25180ed03765c39b78