Improve CarPropertyManagerTest.
This CL improves CarPropertyManager test to cover more code. Originally
some code path are skipped because the mock vhal generates exceptions
that are passed by binder to the caller and directly trigger the
caller's exception handler. However, in a real hal, the error code
is returned via error code, not exceptions, so the logic to handle
error code was not actually covered.

This CL also also add a test case to test null property case.

Test: atest CarServiceTest:com.android.car.CarPropertyManagerTest
on gcar_emu_x86-userdebug target.
Bug: 190231228

Change-Id: If9175da441e5a44bb0c9e9b5f7d1526adf40f599