Make connections more robust immediately after pairing
This is done by inferring whether a device failed a connection attempt,
and retrying the connection if the device hasn't connected before since
bonding. This does not persist across Bluetooth sessions, e.g., if a
device bonds and Bluetooth is restarted before the device successfully
connects on this profile, this will *not* help in the new Bluetooth
session.

Bug: 161910187
Test: atest BluetoothConnectionRetryManagerTest
Test: atest CarBluetoothServiceTest
Test: Manual -- When paring an iPhone running iOS 14.5+, wait ~10
seconds when prompted to allow sharing contacts before accepting. Logcats
should show PBAP failed to connect once, automatic retry, and then
successfully connect. Similarly, enable "Show Notifications" within 20
seconds (in the detailed BT settings on iPhone), and MAP should
auto retry and connect as well.

Change-Id: Idbb4c6ecc61ba5e8de31a5189d9e43c760bd3806