Allow only a single stream client to exist
Because CarEvsService provides the video stream only to the foreground
activity, it would not have more than a single stream client at a time.
Therefore, this CL removes RemoteCallbackList usage and replaces it with
a single ICarEvsStreamCallback member variable.

Also, this adds lines to declien a request to stop the video stream not
from current stream client.

Fix: 178741919
Fix: 182299098
Fix: 183672792
Test: Run CarEvsCameraPreviewApp by injecting events.
Change-Id: I4511f6a6b823d0e5e89cd220ffa5f6ad47237a9d