Add EvsManager null check to tearDown()
If CarEvsService is not enable on the device under test, it breaks an
assumption in CarEvsManagerTest::setUp() and therefore the test executor
calls tearDown() method with a null EvsManager reference.
CarEvsManagerTest::tearDown() requests EvsManager to stop a video stream
and fails to invoke a method because of null EvsManager reference.

To fix this issue, this CL add a null check before invoking a
CarEvsManager method in tearDown().

Fix: 191192986
Test: Run 'atest CarEvsManagerTest -c' and confirm all test results are
      assumption failed on Seahawk.
Change-Id: Ib21e13ebfd646a65e8e207d3f809674a6f96b6c2