Execute scripts based on the publication of data or a signal from system monitor
This CL contains CarTelemetryService-internal implementation and
threading model. We will move all cartelemetry operation on the same
thread.

The following changes are included:
(1) use a priority number instead of enabling a set of publishers
(2) a priority queue of ScriptExecutionTask (containing a reference
to a subscriber and data) is created. When a certain event happens,
the task will be scheduled and the pq gets polled.
(3) event driven task scheduling using a handler
(4) move the add and remove metrics config logic to handler thread

Bug: 190862594
Test: m CarService
Test: atest CarServiceUnitTest:DataBrokerUnitTest

Change-Id: I45e7d9768370f7ce803248a254713128bfd3f51d