[ResultStore] create flat file result store class
This CL creates a flat file result store class. Two folders are created
for interim and final results. The interim results are loaded into memory
for faster access, and they are persisted to disk with the shutdown()
method call. Final results are not stored in cache and are always
immediately persisted/retrieved from disk.

Bug: 195422410
Test: CarServiceUnitTest:ResultStoreTest

Change-Id: Iea103513fd7f932112e1e081b60250878a4f15af