Fix GuardedBy warnings in com.android.car.watchdog
Errorprone offers enforcement of @GuardedBy annotations.
By default they are errors, but until this point have
been downgraded to warnings. Soon they will be made
errors on presubmit, blocking further cls.

This is one cl of many, one is being made for each
package of CarService.

This doesn't fix the issues in WatchdogPerfHandler,
as it's usage of nested classes accessing the outer
class's lock made it tricker to fix, so I skipped
it for now.

Bug: 190944875
Test: m RUN_ERROR_PRONE=true CarServiceUpdatable
Change-Id: Ic523809d60faaed3218edc3fd933a175f8c81b7a