Remove dependency on libbinder by native Script Executor library.
This is a part of the effort to make ScriptExecutor compile against SDK.
The only reason libscriptexecutor depends on binder is because it uses
constant definitions in AIDL file for error codes returned by onError
callback. At the cost of redefining a couple of constants, we gain more
lightweight library and make the task of building against SDK easier.

Fixes: 194324369
Bug: 197006437
Test: atest ScriptExecutorUnitTest:ScriptExecutorTest
Test: atest ScriptExecutorUnitTest:JniUtilsTest
Change-Id: I628a12a32210a8342aa865ecd543364173025871