[RESTRICT AUTOMERGE] Check both MetricsConfig version for add and remove APIs
In add/remove config API, the validity of a MetricsConfig is checked by
MetricsConfigStore. If the MetricsConfigStore returns a success signal
to CarTelemetryService, the add/remove operations will be applied on the
rest of CarTelemetryService (e.g. ResultStore and DataBroker)

This CL also refactor the internal APIs to make sure their signatures
are consistent.

Bug: 198792767
Test: atest CarServiceUnitTest:DataBrokerTest
Test: atest CarServiceUnitTest:DataBrokerControllerTest
Test: atest CarServiceUnitTest:MetricsConfigStoreTest
Test: atest CarServiceUnitTest:ResultStoreTest
Test: atest CarServiceUnitTest:CarTelemetryServiceTest
Test: atest CarServiceUnitTest:CarTelemetryManagerTest

Change-Id: Ia2a0244ef5f876e15e80ee0b5134891808ab5cf7