Fix cyclic dependency in CarPowerPolicyServer.
Refactor CarPowerPolicyServer. All members of CarPowerPolicyServer
now only have a reference to this and are guaranteed to live
shorter then its parent by adding clean-ups in terminate function.

Move CarServiceNotificationHandler instance to be owned by main
function instead of CarServiceNotificationHandler. It now takes
a weak reference to CarPowerPolicyServer so that its lifetime is
not bound with CarPowerPolicyServer. Main function now owns all
the binder objects that would be added as service.

Test: atest carpowerpolicyserver_test
Bug: 210751936
Change-Id: Idd97d12480a19cd91cb54f252d033d7953b7df8b