Fixed flaky test in CarPowerManagerUnitTest
CarPowerManagerUnitTest#testSetListener was flaky
because of incorrect number of events was expected.
Between function return and check for events in the
list, new event was received and added to the list.
As result test case was mostly passing, but ocasionally,
event wasn't recevied before the check.
Fixed number of expected events and changed await() to
return copy of the list, so event cannot be added later.

Bug: 213908746

Test: atest --rerun-until-failure 100 CarPowerManagerUnitTest
Change-Id: I33f5d99eb6e0f21f2cc3999f6d38e951a2da10a7