Manual cherry pick due to refactorings and differences in CarPowerManager API in S from ag/16852155. Also contains all changes in ag/16853938, ag/16873606.
(1) Adds initSession() method to SessionController.

We need to properly handle a case when CarPowerManagementService's
power state is already ON by the time the system boots up.
In such case we want to trigger logic that checks if a new session
signal needs to be generated.
We invoke the initSession method after boot signal was received inside
DataBrokerController to ensure that it gets called after all publishers
were initiated for all of the currently active configs and metric
collection has already started.

(2) Add boot reason to session annotations.

The goal is to propagate information that a crash happened to clients.
For example later when the data is actually analyzed, it will be possible to apply extra analysis if a crash
happened and this collected data needs to be treated differently.

(3) Add conversion of SessionAnnotation to PersistableBundle.

This logic will be used to annotate data that comes in PersistableBundles
from publishers before it gets pushed to subscribers.

Bug: 205611531
Test: atest com.android.car.telemetry
Merged-In: I6265a847c1b8f86b677f0159c3857c5eba1ef3fb
(cherry picked from commit 18ce2489055555464c00d4c6ea87636e5e2c0d66)
Change-Id: Ie54513cbf166f599dc1d669dbd495378fa49135b