Adds initSession() method to SessionController.
We need to properly handle a case when CarPowerManagementService's power state is already ON by the time the system boots up.
In such case we want to trigger logic that checks if a new session
signal needs to be generated.
We invoke the initSession method after boot signal was received inside
DataBrokerController to ensure that it gets called after all publishers
were initiated for all of the currently active configs and metric
collection has already started.

Bug: 205611531
Test: atest com.android.car.telemetry
Change-Id: I6265a847c1b8f86b677f0159c3857c5eba1ef3fb