Fix unregisterListener bug in CarPropertyService.
There is a bug in CarPropertyService that might cause property
unsubscribed from PropertyHalService when unregisterListener
is called even when there are other listeners for the property.

This bug is due to using upateMaxRate == 0 to check whether there
are other clients. However, sampleRate might be 0 for on_change
event so even if updateMaxRate is 0, it doesn't mean the client
array is empty.

Test: Add unit test in CarPropertyServiceTest
Bug: 229327788
Change-Id: Iff5d18a2f5d05ea9be9b8af28908026306a7a99e