Modify immersive request foreground detection logic
For immersive mode, it is desired to ignore requests from applications
that are not currently on the foreground display area. However, the
previous logic of keeping track of these packages would not handle this
properly if we switched from an immersive to a new non-immersive
application since the WindowManager callback would execute prior to the
foreground array update. Therefore, the new logic will allow all
requests as long as it's not in a non-foreground category.

Bug: 235366866
Test: manual
Change-Id: I79f795e431e01a1d1737a1ae7884d376cb3df6bf