Update Railway ref app fixing bugs
Updating the Railway reference app to include the normal permission
QUERY_ALL_PACKAGES which will allow it to view package info for
installed packages. This is needed to get the package UIDs to be used by
the reference app when calling UID based APIs which restrict an ethernet
network to "allowed" packages.

Also fixing the ref app so that the connect network button works even if
the reference app does not have the restricted network permission. This
allows for testing of the connect network button with the reference app
(as oppposed to giving the reference app the restricted network
permission in which case it would always be able to connect to all
networks).

Also fixed issue where view models were crashing on create due to recent
changes in the androidx stack.

Bug: 236976577
Bug: 236394145
Bug: 239643425
Test: manually tested in the app
Change-Id: I96435e3ac0533ec65fe03cf84b49056ed8cb3379