Fixes a state transition in onStop()
When a new activity is launched, CarEvsCameraPreviewActivity will become
the background activity and, however, likely still visible to the users
due to its window type, TYPE_VOLUME_OVERLAY, which is a relatively
higher z-order.  Therefore, the camera preview should not be stopped in
onStop() if CarEvsCameraPreviewActivity is using the system window.

Fix: 238364260
Test: Run a test sequence described in a bug.
Change-Id: Id6bca6655d80b5a7cb621d6971745d1d9bd8be72